+++ tests/meltano/api/controllers/test_sql_controller.py
@@ -40,6 +40,8 @@ class TestSqlController:
         self.assert_join_graph_two_dependency(post)
         self.assert_join_graph_derived_dependency(post)
         self.assert_no_join(post)
+        self.assert_filters_no_join_query(post)
+        self.assert_filters_hda_query(post)
 
     def assert_empty_query(self, post):
         """with no columns no query should be generated"""
@@ -147,8 +149,10 @@ class TestSqlController:
         res = post(payload)
         assert res.status_code == 200, res.data
         assertIsSQL(res.json["sql"])
-        assert 'EXTRACT(\'Week\' FROM "entry"."from") "from.week"' in res.json["sql"]
-        assert re.search(r'GROUP BY.*"from.week"', res.json["sql"])
+        assert (
+            'EXTRACT(\'WEEK\' FROM "entry"."from") "entry.from.week"' in res.json["sql"]
+        )
+        assert re.search(r'GROUP BY.*"entry.from.week"', res.json["sql"])
 
     def assert_join_graph_no_dependencies(self, post):
         payload = {
@@ -166,10 +170,12 @@ class TestSqlController:
         res = post(payload)
 
         assert res.status_code == 200
-        assert (
-            'SELECT "region"."dnoregion" "region.dnoregion" FROM "region" "region" GROUP BY "region.dnoregion" LIMIT 3;'
-            in res.json["sql"]
-        )
+        assertIsSQL(res.json["sql"])
+
+        assert 'SELECT "region"."dnoregion" "region.dnoregion"' in res.json["sql"]
+        assert 'FROM "region" "region"' in res.json["sql"]
+        assert 'GROUP BY "region.dnoregion"' in res.json["sql"]
+        assert 'ORDER BY "region.dnoregion" ASC' in res.json["sql"]
 
     def assert_join_graph_one_dependency(self, post):
         payload = {
@@ -187,8 +193,19 @@ class TestSqlController:
         res = post(payload)
 
         assert res.status_code == 200
-        sql = 'SELECT "region"."dnoregion" "region.dnoregion","entry"."forecast" "entry.forecast" FROM "region" "region" JOIN "entry" "entry" ON "region"."id"="entry"."region_id" GROUP BY "region.dnoregion","entry.forecast" LIMIT 3;'
-        assert sql in res.json["sql"]
+        assertIsSQL(res.json["sql"])
+
+        assert (
+            'SELECT "region"."dnoregion" "region.dnoregion","entry"."forecast" "entry.forecast"'
+            in res.json["sql"]
+        )
+        assert 'FROM "region" "region"' in res.json["sql"]
+        assert (
+            'JOIN "entry" "entry" ON "region"."id"="entry"."region_id"'
+            in res.json["sql"]
+        )
+        assert 'GROUP BY "region.dnoregion","entry.forecast"' in res.json["sql"]
+        assert 'ORDER BY "region.dnoregion" ASC,"entry.forecast" ASC' in res.json["sql"]
 
     def assert_join_graph_two_dependency(self, post):
         payload = {
@@ -209,8 +226,25 @@ class TestSqlController:
         res = post(payload)
 
         assert res.status_code == 200
-        sql = 'SELECT "region"."dnoregion" "region.dnoregion","entry"."forecast" "entry.forecast","generationmix"."fuel" "generationmix.fuel","generationmix"."perc" "generationmix.perc" FROM "region" "region" JOIN "entry" "entry" ON "region"."id"="entry"."region_id" JOIN "generationmix" "generationmix" ON "entry"."id"="generationmix"."entry_id" GROUP BY "region.dnoregion","entry.forecast","generationmix.fuel","generationmix.perc" LIMIT 3;'
-        assert sql in res.json["sql"]
+        assertIsSQL(res.json["sql"])
+
+        assert (
+            'SELECT "region"."dnoregion" "region.dnoregion","entry"."forecast" "entry.forecast","generationmix"."perc" "generationmix.perc","generationmix"."fuel" "generationmix.fuel"'
+            in res.json["sql"]
+        )
+        assert 'FROM "region" "region"' in res.json["sql"]
+        assert (
+            'JOIN "entry" "entry" ON "region"."id"="entry"."region_id" JOIN "generationmix" "generationmix" ON "entry"."id"="generationmix"."entry_id"'
+            in res.json["sql"]
+        )
+        assert (
+            'GROUP BY "region.dnoregion","entry.forecast","generationmix.perc","generationmix.fuel"'
+            in res.json["sql"]
+        )
+        assert (
+            'ORDER BY "region.dnoregion" ASC,"entry.forecast" ASC,"generationmix.perc" ASC,"generationmix.fuel" ASC'
+            in res.json["sql"]
+        )
 
     def assert_join_graph_derived_dependency(self, post):
         payload = {
@@ -231,8 +265,25 @@ class TestSqlController:
         res = post(payload)
 
         assert res.status_code == 200
-        sql = 'SELECT "region"."dnoregion" "region.dnoregion","generationmix"."fuel" "generationmix.fuel","generationmix"."perc" "generationmix.perc" FROM "region" "region" JOIN "entry" "entry" ON "region"."id"="entry"."region_id" JOIN "generationmix" "generationmix" ON "entry"."id"="generationmix"."entry_id" GROUP BY "region.dnoregion","generationmix.fuel","generationmix.perc" LIMIT 3;'
-        assert sql in res.json["sql"]
+        assertIsSQL(res.json["sql"])
+
+        assert (
+            'SELECT "region"."dnoregion" "region.dnoregion","generationmix"."perc" "generationmix.perc","generationmix"."fuel" "generationmix.fuel"'
+            in res.json["sql"]
+        )
+        assert 'FROM "region" "region"' in res.json["sql"]
+        assert (
+            'JOIN "entry" "entry" ON "region"."id"="entry"."region_id" JOIN "generationmix" "generationmix" ON "entry"."id"="generationmix"."entry_id"'
+            in res.json["sql"]
+        )
+        assert (
+            'GROUP BY "region.dnoregion","generationmix.perc","generationmix.fuel"'
+            in res.json["sql"]
+        )
+        assert (
+            'ORDER BY "region.dnoregion" ASC,"generationmix.perc" ASC,"generationmix.fuel" ASC'
+            in res.json["sql"]
+        )
 
     def assert_no_join(self, post):
         payload = {
@@ -250,5 +301,136 @@ class TestSqlController:
         res = post(payload)
 
         assert res.status_code == 200
-        sql = 'SELECT "region"."dnoregion" "region.dnoregion" FROM "region" "region" GROUP BY "region.dnoregion" LIMIT 3;'
+        assertIsSQL(res.json["sql"])
+
+        sql = 'SELECT "region"."dnoregion" "region.dnoregion" FROM "region" "region" GROUP BY "region.dnoregion" ORDER BY "region.dnoregion" ASC LIMIT 3;'
         assert sql in res.json["sql"]
+
+    def assert_filters_no_join_query(self, post):
+        payload = {
+            "table": "region",
+            "columns": ["name", "short_name"],
+            "aggregates": ["count"],
+            "timeframes": [],
+            "joins": [],
+            "order": None,
+            "limit": 3,
+            "filters": {
+                "columns": [
+                    {
+                        "table_name": "region",
+                        "name": "name",
+                        "expression": "equal_to",
+                        "value": "East of England",
+                    },
+                    {
+                        "table_name": "region",
+                        "name": "short_name",
+                        "expression": "like",
+                        "value": "%East%",
+                    },
+                ],
+                "aggregates": [
+                    {
+                        "table_name": "region",
+                        "name": "count",
+                        "expression": "greater_than",
+                        "value": 50,
+                    }
+                ],
+            },
+            "run": False,
+        }
+
+        res = post(payload)
+
+        assert res.status_code == 200
+        assertIsSQL(res.json["sql"])
+
+        assert (
+            'SELECT "region"."dnoregion" "region.dnoregion","region"."shortname" "region.shortname",COALESCE(COUNT("region"."id"),0) "region.count"'
+            in res.json["sql"]
+        )
+        assert 'FROM "region" "region"' in res.json["sql"]
+        assert (
+            'WHERE "region"."dnoregion"=\'East of England\' AND "region"."shortname" LIKE \'%East%\''
+            in res.json["sql"]
+        )
+        assert 'GROUP BY "region.dnoregion","region.shortname"' in res.json["sql"]
+        assert 'HAVING COALESCE(COUNT("region"."id"),0)>50 ' in res.json["sql"]
+        assert (
+            'ORDER BY "region.dnoregion" ASC,"region.shortname" ASC' in res.json["sql"]
+        )
+
+    def assert_filters_hda_query(self, post):
+        payload = {
+            "table": "region",
+            "columns": ["name", "short_name"],
+            "aggregates": ["count"],
+            "timeframes": [],
+            "joins": [
+                {"name": "entry", "columns": ["forecast"]},
+                {
+                    "name": "generationmix",
+                    "columns": ["perc", "fuel"],
+                    "aggregates": ["avg_perc"],
+                },
+            ],
+            "order": None,
+            "limit": 3,
+            "filters": {
+                "columns": [
+                    {
+                        "table_name": "region",
+                        "name": "name",
+                        "expression": "less_than",
+                        "value": "L",
+                    },
+                    {
+                        "table_name": "entry",
+                        "name": "forecast",
+                        "expression": "is_null",
+                        "value": "",
+                    },
+                    {
+                        "table_name": "generationmix",
+                        "name": "fuel",
+                        "expression": "is_not_null",
+                        "value": "",
+                    },
+                    {
+                        "table_name": "generationmix",
+                        "name": "perc",
+                        "expression": "greater_or_equal_than",
+                        "value": "10",
+                    },
+                ],
+                "aggregates": [
+                    {
+                        "table_name": "region",
+                        "name": "count",
+                        "expression": "less_or_equal_than",
+                        "value": 50,
+                    },
+                    {
+                        "table_name": "generationmix",
+                        "name": "avg_perc",
+                        "expression": "greater_than",
+                        "value": 15,
+                    },
+                ],
+            },
+            "run": False,
+        }
+
+        res = post(payload)
+
+        assert res.status_code == 200
+        assertIsSQL(res.json["sql"])
+
+        assert (
+            'WHERE "region"."dnoregion"<\'L\' AND "entry"."forecast" IS NULL AND "generationmix"."perc">=\'10\' AND NOT "generationmix"."fuel" IS NULL'
+            in res.json["sql"]
+        )
+        assert 'HAVING COALESCE(COUNT("region.id"),0)<=50' in res.json["sql"]
+        assert 'HAVING COALESCE(AVG("generationmix.perc"),0)>15' in res.json["sql"]